requirements.txt
djangorestframework>=1.11
validators>=0.20.0
xmltodict>=0.13.0
